+++ admin/package.json
@@ -18,6 +18,7 @@
     "@browser-logos/vivaldi": "^2.1.10",
     "@browser-logos/webkit": "^1.0.9",
     "@egoistdeveloper/operating-system-logos": "^1.2.3",
+    "@img-comparison-slider/react": "^8.0.0",
     "@tanstack/query-async-storage-persister": "^4.24.4",
     "@tanstack/query-sync-storage-persister": "^4.24.4",
     "@tanstack/react-query-devtools": "^4.24.3",
