+++ admin/src/components/TablePanels.jsx
@@ -8,7 +8,7 @@ import ImportPanel from './ImportPanel';
 import DangerPanel from './DangerPanel';
 import DetailsPanel from './DetailsPanel';
 import { ReactComponent as Trash } from '../assets/images/icons/icon-trash.svg';
-import ChangesPanel from './ChangesPanel';
+import ChangesPanel from './ChangesPanel/ChangesPanel';
 
 export default function TablePanels( { props } ) {
 	const { options, initialRow, handlePanel } = props;
