+++ admin/src/hooks/useTablePanels.jsx
@@ -2,6 +2,7 @@ import { create } from 'zustand';
 
 const useTablePanels = create( ( set ) => ( {
 	activePanel: undefined,
+	imageCompare: false,
 	options: [],
 	rowToEdit: {},
 	activatePanel: ( activePanel ) => set( () => ( { activePanel } ) ),
